Rename 'packaging' package to 'cli'
This is more suiting name for a package which module's common property
is usage from CLI rather than in application runtime.
